Moved checks into better hierachy and refactored code
