+++ profiles/npro.yml
@@ -2,4 +2,4 @@ metadata :
   machine_name: 'npro'
   title : 'NPro'
 checks :
-  \SiteAudit\Drush\ShieldDisabled : {}
+  \SiteAudit\Check\Drush\ShieldDisabled : {}
