+++ profiles/sabm.yml
@@ -2,7 +2,7 @@ metadata :
   machine_name: 'sabm'
   title : 'SABM'
 checks :
-  \SiteAudit\Check\ModuleEnabled:
+  \SiteAudit\Check\Drush\ModuleEnabled:
     modules:
       - syslog
       - acsf
@@ -13,17 +13,17 @@ checks :
       - acsf_variables
       - acquia_spi
       - acquia_agent
-  \SiteAudit\Check\ModuleDisabled:
+  \SiteAudit\Check\Drush\ModuleDisabled:
     modules:
       - dblog
       - overlay
       - php
       - statistics
-  \SiteAudit\Drush\PreprocessCSS : {}
-  \SiteAudit\Drush\PreprocessJS : {}
-  \SiteAudit\Drush\PageCacheMaximumAge :
+  \SiteAudit\Check\Drush\PreprocessCSS : {}
+  \SiteAudit\Check\Drush\PreprocessJS : {}
+  \SiteAudit\Check\Drush\PageCacheMaximumAge :
     cache : 300
-  \SiteAudit\Drush\UpdateDBStatus: {}
+  \SiteAudit\Check\Drush\UpdateDBStatus: {}
   \SiteAudit\Check\Acsf\ThemeSize :
     max_size : 50
     warning_size: 20
