+++ src/AuditResponse/system/cron_has_run.yml
@@ -1,0 +1,6 @@
+description: "Checks that cron has run recently."
+remediation: "Ensure a cron job has been configured for the site. If so, file a support ticket to investigate why cron has stopped working."
+messages:
+  success: 'Cron was last run on :cron_last'
+  failure: 'Cron was last run on :cron_last'
+  exception: 'Could not determine status of cron: :exception'
