+++ src/Check/Drush/PageCacheMaximumAge.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace SiteAudit\Check\Drush;
+
+use SiteAudit\Check\Check;
+
+class PageCacheMaximumAge extends Check {
+  protected function getNamespace()
+  {
+    return 'variable/pagecache';
+  }
+
+  public function check() {
+    $context = $this->context;
+    $value = $context->drush->getVariable('page_cache_maximum_age', 0);
+
+    if (is_int($value) || is_string($value)) {
+      $this->setToken('value', $value);
+      return ((int) $value) >= $this->getOption('cache', 300);
+    }
+
+    return FALSE;
+  }
+}
